Conflicts:
	brn-conf.sh
	configure
	doc/testie.1

+++ elements/brn2/wifi/wifierrorclassifier.cc
@@ -12,7 +12,12 @@ CLICK_DECLS
 WifiErrorClassifier::WifiErrorClassifier()
   : _p_ok(0),
     _p_crc(0),
-    _p_phy(0)
+    _p_phy(0),
+    _p_fifo(0),
+    _p_decrypt(0),
+    _p_mic(0),
+    _p_zerorate(0),
+    _p_unknown(0)
 {
 }
 
@@ -24,14 +29,22 @@ Packet *
 WifiErrorClassifier::simple_action(Packet *p)
 {
   struct click_wifi_extra *ceha = WIFI_EXTRA_ANNO(p);
-  struct click_wifi_extra *cehp = (struct click_wifi_extra *) p->data();
 
-
-  if ( (ceha->magic == WIFI_EXTRA_MAGIC && ceha->flags & WIFI_EXTRA_RX_ERR) ||
-       (cehp->magic == WIFI_EXTRA_MAGIC && cehp->flags & WIFI_EXTRA_RX_ERR) )
+  if ( (ceha->magic == WIFI_EXTRA_MAGIC && ceha->flags & WIFI_EXTRA_RX_ERR) )
   {
-    if ( (ceha->magic == WIFI_EXTRA_MAGIC && ceha->flags & WIFI_EXTRA_RX_CRC_ERR) ||
-         (cehp->magic == WIFI_EXTRA_MAGIC && cehp->flags & WIFI_EXTRA_RX_CRC_ERR) )
+
+    if ( ceha->flags & WIFI_EXTRA_RX_ZERORATE_ERR )
+    {
+      _p_zerorate++;
+      if ( noutputs() > 6 )
+        output(6).push(p);
+      else
+        p->kill();
+
+      return 0;
+    }
+
+    if ( ceha->flags & WIFI_EXTRA_RX_CRC_ERR )
     {
       _p_crc++;
       if ( noutputs() > 1 )
@@ -42,8 +55,7 @@ WifiErrorClassifier::simple_action(Packet *p)
       return 0;
     }
 
-    if ( (ceha->magic == WIFI_EXTRA_MAGIC && ceha->flags & WIFI_EXTRA_RX_PHY_ERR) ||
-          (cehp->magic == WIFI_EXTRA_MAGIC && cehp->flags & WIFI_EXTRA_RX_PHY_ERR) )
+    if ( ceha->flags & WIFI_EXTRA_RX_PHY_ERR )
     {
       _p_phy++;
       if ( noutputs() > 2 )
@@ -54,7 +66,47 @@ WifiErrorClassifier::simple_action(Packet *p)
       return 0;
     }
 
-    p->kill();
+    if ( ceha->flags & WIFI_EXTRA_RX_FIFO_ERR )
+    {
+      _p_fifo++;
+      if ( noutputs() > 3 )
+        output(3).push(p);
+      else
+        p->kill();
+
+      return 0;
+    }
+
+    if ( ceha->flags & WIFI_EXTRA_RX_DECRYPT_ERR )
+    {
+      _p_decrypt++;
+      if ( noutputs() > 4 )
+        output(4).push(p);
+      else
+        p->kill();
+
+      return 0;
+    }
+
+    if ( ceha->flags & WIFI_EXTRA_RX_MIC_ERR )
+    {
+      _p_mic++;
+      if ( noutputs() > 5 )
+        output(5).push(p);
+      else
+        p->kill();
+
+      return 0;
+    }
+
+    _p_unknown++;
+    if ( noutputs() > 7 )
+      output(7).push(p);
+    else {
+      click_chatter("Discard unknown error");
+      p->kill();
+    }
+
     return 0;
   }
